backend/requirements.txt
browser-use>=0.1.0
playwright>=1.46
python-dotenv>=1.0.1
